Move GetObjects to derived EdgeCollection classes for a stronger typed API.
