+++ src/AsmResolver.Workspaces/Collections/EdgeCollection.cs
@@ -211,30 +211,6 @@ namespace AsmResolver.Workspaces.Collections
         }
 
         /// <summary>
-        /// Gets a collection of all objects that are related to this object of a given relation type.
-        /// </summary>
-        /// <param name="relation">The relation.</param>
-        /// <typeparam name="TSource">The type of the source object.</typeparam>
-        /// <typeparam name="TTarget">The type of object to obtain.</typeparam>
-        public IEnumerable<TTarget> GetObjects<TSource, TTarget>(ObjectRelation<TSource, TTarget> relation)
-        {
-            return GetEdges(relation)
-                .Select(e => (TTarget) GetAdjacentNode(e).Subject)
-                .Distinct();
-        }
-
-        /// <summary>
-        /// Gets a collection of all objects that are related to this object of any of the given relation types.
-        /// </summary>
-        /// <param name="relations">The relations to include in the lookup.</param>
-        public IEnumerable<object> GetObjects(params ObjectRelation[] relations)
-        {
-            return GetEdges(relations)
-                .Select(n => GetAdjacentNode(n).Subject)
-                .Distinct();
-        }
-
-        /// <summary>
         /// Gets a collection of all nodes that are related to this object.
         /// </summary>
         public IEnumerable<object> GetAllObjects()
